Add include for captioned images and corresponding SCSS
Using Liquid markup the image size, url and caption are inserted
when the snippet is included in a document.

SCSS is added for centring of the figures. Wrap the Liquid inlucde
in a div with class="centre-block". The figure snippet uses the
centre-fig class for centreing of images inside the block.